+++ CHANGELOG.md
@@ -7,10 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [1.0.2] - 2022-12-23
+
+### Fixed
+
+- Multiples bugs in audit tools used to dump and verify events
+
+
+## [1.0.1] - 2022-12-19
+
 ### Added
 
 - This CHANGELOG
 - Test to search with dates as filter
+- Functions to get token and domain according to test environment (PROD/DEV/STG)
 
 ### Fixed